Finished changing the package name, this time in all imports, and added the missing object/*.java code to the project.
Changed from "girlsofsteel" to "com.girlsofsteelrobotics.atlas"
throughout all the import statements. (Yes, there was an easier way to
do this in Eclipse, if I had imported the original code in a different
way.) Also fixed up a couple of references to MathUtils for functions
now found in the standard Math library.