Important fix to drive distance calculation code: remove Math.abs() plus cleanups
There were too many Math.abs() calls in the code that calculates the
driving distance. It would produce incorrect results when the starting
and ending encoder values are on opposite sides of zero.

Also fixed up local variable naming to be consistent and consolidated
the throttle to speed conversion code to a private method.